Sanitize names
Added a preg_replace that sanitizes the name. As per the comment, this
is a very basic check as names are very diverse and difficult to sanitize.
Some countries have very specific unicode characters that are used and we need
to be inclusive in allowing these names with special characters.